# docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorBody:
    properties:
      code:
        description: Code is one of the Code* constants; clients branch on it, unlike
          Message.
        example: validation_failed
        type: string
      fields:
        additionalProperties:
          type: string
        description: Fields is set for request-validation failures only.
        type: object
      message:
        description: Message is human-readable and may change freely.
        example: invalid request
        type: string
    type: object
  github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorBody'
      success:
        example: false
        type: boolean
    type: object
  github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        example: true
        type: boolean
    type: object
  internal_athlete.athleteResponse:
    properties:
      athlete_type:
        example: private
        type: string
      created_at:
        example: "2026-09-24T09:30:00Z"
        type: string
      experience_level:
        example: beginner
        type: string
      first_name:
        example: Sara
        type: string
      goal:
        example: lose 5kg
        type: string
      height:
        example: 172
        type: number
      id:
        example: 3f0b1c6e-2a1d-4f7b-9c3e-6d5a4b3c2d1e
        type: string
      injuries:
        example: left shoulder impingement
        type: string
      last_name:
        example: Ahmadi
        type: string
      phone:
        example: "09372144430"
        type: string
      updated_at:
        example: "2026-09-24T09:30:00Z"
        type: string
    type: object
  internal_athlete.createRequest:
    properties:
      athlete_type:
        example: private
        type: string
      experience_level:
        example: beginner
        type: string
      first_name:
        example: Sara
        type: string
      goal:
        example: lose 5kg
        type: string
      height:
        example: 172
        type: number
      injuries:
        example: left shoulder impingement
        type: string
      last_name:
        example: Ahmadi
        type: string
      phone:
        example: "09372144430"
        type: string
    type: object
  internal_athlete.listMeta:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 42
        type: integer
    type: object
  internal_coach.coachResponse:
    properties:
      first_name:
        example: Ada
        type: string
      id:
        example: b8f1c9de-8f0a-4c1e-9a2b-2f6b6b8e2b3a
        type: string
      last_name:
        example: Lovelace
        type: string
      phone:
        example: "09372144430"
        type: string
    type: object
  internal_coach.loginRequest:
    properties:
      password:
        example: correct-horse-battery-staple
        type: string
      phone:
        example: "09372144430"
        type: string
    required:
    - password
    - phone
    type: object
  internal_coach.loginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_coach.signupRequest:
    properties:
      first_name:
        example: Ada
        type: string
      last_name:
        example: Lovelace
        type: string
      password:
        example: correct-horse-battery-staple
        minLength: 8
        type: string
      phone:
        example: "09372144430"
        type: string
    required:
    - first_name
    - last_name
    - password
    - phone
    type: object
  internal_measurement.createRequest:
    properties:
      arm:
        example: 34
        type: number
      chest:
        example: 98
        type: number
      date:
        example: "2026-09-24"
        type: string
      hip:
        example: 99
        type: number
      thigh:
        example: 56
        type: number
      waist:
        example: 81.5
        type: number
      weight:
        example: 72.5
        type: number
    type: object
  internal_measurement.measurementResponse:
    properties:
      arm:
        example: 34
        type: number
      athlete_id:
        example: 3f0b1c6e-2a1d-4f7b-9c3e-6d5a4b3c2d1e
        type: string
      chest:
        example: 98
        type: number
      created_at:
        example: "2026-09-24T09:30:00Z"
        type: string
      date:
        example: "2026-09-24"
        type: string
      hip:
        example: 99
        type: number
      id:
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        type: string
      thigh:
        example: 56
        type: number
      updated_at:
        example: "2026-09-24T09:30:00Z"
        type: string
      waist:
        example: 81.5
        type: number
      weight:
        example: 72.5
        type: number
    type: object
  internal_server.healthzErrorResponse:
    properties:
      reason:
        example: failed to connect to database
        type: string
      status:
        example: unavailable
        type: string
    type: object
  internal_server.healthzOKResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: API for the GoGym coaching platform.
  title: GoGym API
  version: "1.0"
paths:
  /api/v1/athletes:
    get:
      description: 'Lists the authenticated coach''s athletes, newest first. Defaults
        to private athletes only — the coach''s ongoing working list — since public
        athletes are one-off plan-link deliveries; pass athlete_type to widen that.
        Paginated: meta carries the total, limit and offset.'
      parameters:
      - description: Case-insensitive partial match on first or last name
        in: query
        name: q
        type: string
      - description: Which athletes to include; defaults to private
        enum:
        - private
        - public
        - all
        in: query
        name: athlete_type
        type: string
      - default: 20
        description: Page size, 1-100
        in: query
        name: limit
        type: integer
      - default: 0
        description: Rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_athlete.athleteResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_athlete.listMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List athletes
      tags:
      - athletes
    post:
      consumes:
      - application/json
      description: Adds an athlete to the authenticated coach's roster. The owning
        coach comes from the bearer token — a coach_id in the body is ignored. Omitting
        athlete_type creates a private athlete.
      parameters:
      - description: Athlete details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_athlete.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_athlete.athleteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create an athlete
      tags:
      - athletes
  /api/v1/athletes/{id}:
    get:
      description: Returns the athlete's full profile. An athlete that belongs to
        another coach is reported as not found, the same as one that doesn't exist.
      parameters:
      - description: Athlete ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_athlete.athleteResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get an athlete
      tags:
      - athletes
  /api/v1/athletes/{id}/measurements:
    get:
      description: 'Returns every measurement for the athlete, oldest first, ready
        to plot as a trend chart. Not paginated: one check-in per visit stays small.
        An athlete with no measurements yields an empty array; an athlete that belongs
        to another coach is reported as not found.'
      parameters:
      - description: Athlete ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_measurement.measurementResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List measurements
      tags:
      - measurements
    post:
      consumes:
      - application/json
      description: Records one check-in for the athlete. date is required (YYYY-MM-DD);
        every metric is optional, but at least one must be given. Weight is in kilograms,
        circumferences in centimetres. An athlete that belongs to another coach is
        reported as not found.
      parameters:
      - description: Athlete ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Measurement values
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_measurement.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_measurement.measurementResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a measurement
      tags:
      - measurements
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Verifies phone + password and returns a signed JWT. The error is
        the same whether the phone or the password was wrong.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_coach.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_coach.loginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      summary: Log in a coach
      tags:
      - auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a coach account. v1 is single-coach, so this is open with
        no admin gate.
      parameters:
      - description: Signup details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_coach.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_coach.coachResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      summary: Sign up a coach
      tags:
      - auth
  /api/v1/coaches/me:
    get:
      description: Returns the profile of the coach identified by the bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_coach.coachResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AbolfazlTayyar_gogym_internal_httpx.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get the authenticated coach
      tags:
      - auth
  /healthz:
    get:
      description: Reports whether the service and its database connection are healthy.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.healthzOKResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server.healthzErrorResponse'
      summary: Health check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"